% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buckets.R
\name{add_label_to_bucket}
\alias{add_label_to_bucket}
\title{Associate an influx labelID with an influx bucketID}
\usage{
add_label_to_bucket(con, labelID, bucketID)
}
\arguments{
\item{con}{influxdb connection previously established with \code{connect}}

\item{labelID}{id of the label to associate}

\item{bucketID}{id of the label to associate}
}
\value{
Influxdb V2 API json response
}
\description{
Associate an influx labelID with an influx bucketID
}
