% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{add_label_to_variable}
\alias{add_label_to_variable}
\title{Associate an influx labelID with an influx variableID}
\usage{
add_label_to_variable(con, labelID, variableID)
}
\arguments{
\item{con}{the connection previously established with \code{connect}}

\item{labelID}{id of the label to associate}

\item{variableID}{id of the label to associate}
}
\value{
Influx V2 API \code{httr::GET} response
}
\description{
Associate an influx labelID with an influx variableID
}
