% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buckets.R
\name{add_member_to_bucket}
\alias{add_member_to_bucket}
\title{Associate an influx userID with an influx bucketID}
\usage{
add_member_to_bucket(con, id, bucketID, name = NULL)
}
\arguments{
\item{con}{influxdb connection previously established with \code{connect}}

\item{id}{id of the member to associate}

\item{bucketID}{id of the bucket to associate}

\item{name}{optional name of the member to associate}
}
\value{
Influxdb V2 API json response
}
\description{
Associate an influx userID with an influx bucketID
}
