% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorizations.R
\name{all_permissions}
\alias{all_permissions}
\title{Create a set of read/write permissions for all resources}
\usage{
all_permissions()
}
\value{
A permissions data structure that enables read/write access to all resources
}
\description{
Create a set of read/write permissions for all resources
}
\examples{
all_permissions()
}
