% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorizations.R
\name{all_read_permissions}
\alias{all_read_permissions}
\title{Create a set of read permissions for all resources}
\usage{
all_read_permissions()
}
\value{
A permissions data structure that enables read access to all resources
}
\description{
Create a set of read permissions for all resources
}
\examples{
all_read_permissions()
}
