% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{api_path}
\alias{api_path}
\title{Build an Influxdb 2.X API endpoint path}
\usage{
api_path(path, ...)
}
\arguments{
\item{path}{Path to append to the API endpoint}

\item{...}{Additional elements to append to the path separated by /}
}
\value{
All inputs concatenated into a single / separated string
}
\description{
Build an Influxdb 2.X API endpoint path
}
