% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{Create a connection to an influxdb database}
\usage{
connect(
  url = getOption("influxr.url", NULL),
  org = getOption("influxr.org", NULL),
  token = getOption("influxr.token", NULL)
)
}
\arguments{
\item{url}{The influxdb server URL}

\item{org}{The influxdb organization name to use when accessing the server}

\item{token}{The influxdb user authorization token to use when accessing the
server}
}
\value{
an influxdb s3 object that maintains the connection information
}
\description{
The \code{connect()} function validates that the url, org, and token are
valid and returns an s3 object containing that information.  The returned
connection object is the first argument of all other API functions and tells
those functions with which server to communicate. Users are not required to
close connections and an arbitrary number of connections can be established.
}
\examples{
\dontrun{
con <- connect(url = "https://myserver.mycompany.com:8086",
               org = "my-org",
               token = "my-token")
}

}
