% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ConstantVariableProperties}
\alias{ConstantVariableProperties}
\title{Generate a ConstantVariableProperties variable type from a vector or list}
\usage{
ConstantVariableProperties(values)
}
\arguments{
\item{values}{Either a vector or list of property values}
}
\value{
A data structure in the format required by \code{create_variable} to create
a MapVariableProperties variable
}
\description{
Constant variable properties are stored in influx as strings.  Therefore, all elements in the
\code{values} argument should be of length 1 and should be capable of being coerced to a
string if not already a string.
}
\examples{
ConstantVariableProperties(1)
ConstantVariableProperties("ok")
ConstantVariableProperties(c("value1", "value2"))
ConstantVariableProperties(list("value1", "value2"))

}
