% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{convert_to_line_protocol}
\alias{convert_to_line_protocol}
\title{Convert a dataframe to line protocol}
\usage{
convert_to_line_protocol(
  data,
  measurement = NULL,
  measurement_col = NULL,
  time_col,
  tag_cols,
  field_cols = NULL
)
}
\arguments{
\item{data}{the dataframe containing the data to convert to line protocol format}

\item{measurement}{the name of the influx measurement to use.  Cannot be specified if
\code{measurement_col} is also specified.}

\item{measurement_col}{the name of the column in the data frame that contains the measurement
name to use for each row of data.  Cannot be specified if \code{measurement} is also specified.}

\item{time_col}{The name of the column containing the timepoint for each row.}

\item{tag_cols}{The names of the columns that are to be treated as tags by influx}

\item{field_cols}{The names of the columns that are to be treated as fields by influx}
}
\value{
character string of line protocol lines
}
\description{
Convert a dataframe to line protocol
}
