% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorizations.R
\name{create_authorization}
\alias{create_authorization}
\title{Create an influx authorization}
\usage{
create_authorization(
  con,
  description = NULL,
  orgID = NULL,
  permissions = all_read_permissions(),
  status = c("active", "inactive"),
  userID = NULL
)
}
\arguments{
\item{con}{influxdb connection previously established with \code{connect}}

\item{description}{description of the authorization}

\item{orgID}{ID of org that authorization is scoped to}

\item{permissions}{vector of authorized permissions created with \code{permissions} or with
predefined permissions through \code{all_permissions} and \code{all_read_permissions}}

\item{status}{if inactive the token is inactive and requests using the token will be rejected}

\item{userID}{ID of user that authorization is scoped to}
}
\value{
Influx V2 API json response
}
\description{
Create an influx authorization
}
