% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buckets.R
\name{create_bucket}
\alias{create_bucket}
\title{Create a bucket on the influx server}
\usage{
create_bucket(
  con,
  name,
  orgID = NULL,
  retentionRules = retentionRules_(),
  description = NULL,
  rp = NULL,
  schemaType = NULL
)
}
\arguments{
\item{con}{influxdb connection previously established with \code{connect}}

\item{name}{name of the bucket}

\item{orgID}{organization ID}

\item{retentionRules}{influx retention rules for the bucket}

\item{description}{description of the bucket}

\item{rp}{undefined string - TBD}

\item{schemaType}{"implicit" or "explicit"}
}
\value{
Influxdb V2 API json response
}
\description{
Create a bucket on the influx server
}
