% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{create_label}
\alias{create_label}
\title{Create a label on the server running at the influx connection}
\usage{
create_label(con, name, properties = NULL, orgID = NULL)
}
\arguments{
\item{con}{influxdb connection previously established with \code{connect}}

\item{name}{name of the label}

\item{properties}{Key value pairs to store in the label.  Should be the output
of the \code{label_properties} function}

\item{orgID}{ID of organization the label is scoped to.  Defaults to orgID of \code{con}}
}
\value{
influxdb V2 API json response
}
\description{
Create a label on the server running at the influx connection
}
\examples{
\dontrun{
create_label(con, name = "label_name", properties = label_properties(list(a = 1, b = 2)))
}
}
