% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations.R
\name{create_organization}
\alias{create_organization}
\title{Create a organization on the influxdb server}
\usage{
create_organization(con, name, description = NULL)
}
\arguments{
\item{con}{influxdb connection previously established with \code{connect()}}

\item{name}{name of the organization}

\item{description}{description of the organization}
}
\value{
Influxdb V2 API json response
}
\description{
Create a organization on the influxdb server
}
