% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{create_user}
\alias{create_user}
\title{Create a new influxdb user}
\usage{
create_user(con, name, oauthID = NULL, status = c("active", "inactive"))
}
\arguments{
\item{con}{the influx connection previously established with \code{connect}}

\item{name}{name of the user}

\item{oauthID}{string}

\item{status}{either "active" or "inactive"}
}
\value{
Influx V2 API \code{httr::POST} response
}
\description{
Create a new influxdb user
}
