% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{create_variable}
\alias{create_variable}
\title{Create a variable on the server running at the influx connection}
\usage{
create_variable(
  con,
  name,
  arguments,
  orgID = NULL,
  createdAt = NULL,
  description = NULL,
  labels = NULL,
  selected = NULL,
  updatedAt = NULL
)
}
\arguments{
\item{con}{the influx connection previously established with \code{connect}}

\item{name}{name of the variable}

\item{arguments}{value of the variable.  It should be the output of either
\code{QueryVariableProperties}, \code{ConstantVariableProperties}, or
\code{MapVariableProperties}}

\item{orgID}{organization id the variable is to be associated with.  The default value is the
orgID from the connection.}

\item{createdAt}{date the parameter was created}

\item{description}{string description of the variable}

\item{labels}{labels Labels to create/attach to the variable. Should be the output of
the \code{variable_labels} function}

\item{selected}{default key to use if the variable store multiple key/value pairs}

\item{updatedAt}{date the parameter was last updated}
}
\value{
Influx V2 API json response
}
\description{
Create a variable on the server running at the influx connection
}
\examples{
\dontrun{
create_variable(tdb, name = "dims", arguments = MapVariableProperties(list(m=5, n=2)))
}
}
