% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorizations.R
\name{delete_authorization}
\alias{delete_authorization}
\title{Update an authorization from the influx database}
\usage{
delete_authorization(con, authID)
}
\arguments{
\item{con}{influxdb connection previously established with \code{connect}}

\item{authID}{influx authorizationID of the authorization to update}
}
\value{
influxdb V2 API response
}
\description{
Update an authorization from the influx database
}
