% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buckets.R
\name{delete_bucket}
\alias{delete_bucket}
\title{Delete a bucket from the influx database}
\usage{
delete_bucket(con, bucketID)
}
\arguments{
\item{con}{influxdb connection previously established with \code{connect}}

\item{bucketID}{influx bucketID of the bucket to delete}
}
\value{
Influxdb V2 API http response
}
\description{
Delete a bucket from the influx database
}
