% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{delete_label}
\alias{delete_label}
\title{Delete a label from the influx database}
\usage{
delete_label(con, labelID)
}
\arguments{
\item{con}{influxdb connection previously established with \code{connect}}

\item{labelID}{labelID of the label to delete}
}
\value{
influxdb V2 API json response
}
\description{
Delete a label from the influx database
}
