% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buckets.R
\name{delete_label_from_bucket}
\alias{delete_label_from_bucket}
\title{Delete the association of an influx labelID with an influx bucketID}
\usage{
delete_label_from_bucket(con, labelID, bucketID)
}
\arguments{
\item{con}{influxdb connection previously established with \code{connect}}

\item{labelID}{id of the label to associate}

\item{bucketID}{id of the label to associate}
}
\value{
Influxdb V2 API json response
}
\description{
Delete the association of an influx labelID with an influx bucketID
}
