% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations.R
\name{delete_organization}
\alias{delete_organization}
\title{Delete a organization from the influxdb database}
\usage{
delete_organization(con, orgID)
}
\arguments{
\item{con}{influxdb connection previously established with \code{connect()}}

\item{orgID}{influx orgID of the organization to delete}
}
\value{
Influxdb V2 API response
}
\description{
Delete a organization from the influxdb database
}
