% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{delete_secret}
\alias{delete_secret}
\title{Delete the association of an influxdb secretID with an influx orgID}
\usage{
delete_secret(con, orgID, secretID)
}
\arguments{
\item{con}{influxdb connection previously established with \code{connect}}

\item{orgID}{id of the organization to disassociate}

\item{secretID}{id of the secrete to delete}
}
\value{
influxdb V2 API http response
}
\description{
Delete the association of an influxdb secretID with an influx orgID
}
