% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{delete_source}
\alias{delete_source}
\title{Delete an influxdb source}
\usage{
delete_source(con, sourceID)
}
\arguments{
\item{con}{influxdb connection previously established with \code{connect}}

\item{sourceID}{string source identifier}
}
\value{
influxdb V2 API http response
}
\description{
Delete an influxdb source
}
