% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{delete_user}
\alias{delete_user}
\title{Delete a user from the influxdb database}
\usage{
delete_user(con, userID)
}
\arguments{
\item{con}{the influx connection previously established with \code{connect}}

\item{userID}{influx userID of the user to delete}
}
\value{
Influx V2 API \code{httr::DELETE} response
}
\description{
Delete a user from the influxdb database
}
