% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{delete_variable}
\alias{delete_variable}
\title{Delete a variable from the influx database}
\usage{
delete_variable(con, variableID)
}
\arguments{
\item{con}{the influx connection previously established with \code{connect}}

\item{variableID}{influx variableID of the variable to delete}
}
\value{
Influx V2 API json response
}
\description{
Delete a variable from the influx database
}
