% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_tibble.R
\name{generate_influx_response_parser}
\alias{generate_influx_response_parser}
\title{Generate a influx stream parser that can be assigned to the callback argument
of httr2::req_stream()}
\usage{
generate_influx_response_parser(output)
}
\arguments{
\item{output}{A InfluxParserOutput object that is used by the parser to pass
information back to the function that called httr2::req_stream().}
}
\value{
A function that takes a single argument containing the raw data read
  from the stream.
}
\description{
httr2::req_stream() requires a function that takes a single argument
which is the raw bytes returned from the stream.  Because the parser state
must be maintained by parser, this parser is created as a closure.
}
