% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{get_source_buckets}
\alias{get_source_buckets}
\title{Get buckets from a source}
\usage{
get_source_buckets(con, sourceID, org = NULL)
}
\arguments{
\item{con}{influxdb connection previously established with \code{connect}}

\item{sourceID}{the source ID.}

\item{org}{The name of the organization}
}
\value{
Influx V2 API json response
}
\description{
Get buckets from a source
}
