% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{get_source_health}
\alias{get_source_health}
\title{Obtain the health information from an influxdb source}
\usage{
get_source_health(con, sourceID)
}
\arguments{
\item{con}{influxdb connection previously established with \code{connect}}

\item{sourceID}{string source identifier}
}
\value{
influxdb V2 API json response
}
\description{
Obtain the health information from an influxdb source
}
