% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health.R
\name{health}
\alias{health}
\title{Retrieve the health status of an instance}
\usage{
health(con)
}
\arguments{
\item{con}{influxdb connection previously established with \code{connect}}
}
\value{
Influxdb API V2.0 json response
}
\description{
Retrieve the health status of an instance
}
