% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{influxdb_class}
\alias{influxdb_class}
\title{Return the class name for an influxdb connection}
\usage{
influxdb_class()
}
\value{
string class name
}
\description{
Return the class name for an influxdb connection
}
\examples{
influxdb_class()
}
