% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorizations.R
\name{influxdb_resource_types}
\alias{influxdb_resource_types}
\title{Return a character vector of all available resource types in influx}
\usage{
influxdb_resource_types()
}
\value{
character vector of all available resource types in influx
}
\description{
Return a character vector of all available resource types in influx
}
\examples{
influxdb_resource_types()
}
