% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_tibble.R
\docType{class}
\name{InfluxParserOutput-class}
\alias{InfluxParserOutput-class}
\alias{InfluxParserOutput}
\title{A reference class that maintains the output of the http stream parser}
\value{
An instance of a InfluxParserOutput object
}
\description{
Anytime httr2::req_stream() is called, it must be supplied the name of
a parser to parse the stream.  In order to maintain output information between
successive calls to the parser, a reference class object is maintained in
the parser function closure.
}
\section{Fields}{

\describe{
\item{\code{tables}}{ANY. The list of CSV tables returned after processing all data
chunks. For most queries, only a single table is returned as the first
element of the list.}

\item{\code{error_message}}{character. The contents of the http body when something
other than a CSV table is received.}

\item{\code{table_chunks}}{ANY. The list of partial CSV tables that are used to form
the current full CSV table being parsed across multiple chunks.}

\item{\code{record_count}}{integer. The total number of CSV records read across all
tables.}

\item{\code{parser_count}}{integer. The number of times the parser was called.  This
is equivalent to the number of http chunks received.}

\item{\code{debug}}{logical. A flag indicating that the parser should output
debug messages.}
}}


