% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{influxr_connection}
\alias{influxr_connection}
\title{Define a connection to InfluxDB instance}
\usage{
influxr_connection(
  host = "localhost",
  port = 8086,
  username = Sys.getenv("INFLUX_USERNAME"),
  password = Sys.getenv("INFLUX_PASSWORD"),
  ssl = TRUE,
  ssl_verifypeer = TRUE,
  timeout = 0,
  retries = 3,
  chunk_size = 20000L,
  post_chunk_size = 50000L,
  server_timezone = "UTC"
)
}
\arguments{
\item{host}{character string. Hostname of InfluxDB server, default is 'localhost'}

\item{port}{numeric. Port number to connect to InfluxDB instance, default is 8086}

\item{username}{character string. User name to use for authentication, default is getting it from the environment
variable INFLUX_USERNAME}

\item{password}{character, password to use for authentication, default is getting it from the environment variable INFLUX_PASSWORD}

\item{ssl}{logical. Whether https should be used instead of http to connect to InfluxDB, defaults to TRUE.}

\item{ssl_verifypeer}{logical. Whether to verify SSL peer.}

\item{timeout}{numeric. Number of seconds requests will wait for the client to
establish a connection, default is 0 (unlimited).}

\item{retries}{numeric. Number of retries the client will try to connect before 
aborting, default is 3, NULL indicates try until success.}

\item{chunk_size}{numeric. Number of data points to fetch from influxDB server per chunk, 
default to 20000.}

\item{post_chunk_size}{numeric. Number of data points uploaded at each chunk when posting data to 
influxDB server.}

\item{server_timezone}{character string. The timezone code of the data in the server. Uploaded data will be converted to it. InfluxDB keeps data in 
\code{'UTC'} - currently only 'UTC' is supported.}
}
\description{
\code{influxr_connection} Creates an `Influx_client` object which contains connection parameters and
 facilitates connecting to an InfluxDB server.
}
\details{
The returned object will hold all the information necessary to connect to InfluxDB and shall be passed to functions
responsible for connecting to the database. Use \code{'summary()'} on the client object to view basic information.
}
