% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fread-recover.R
\name{influxr_fread_recover}
\alias{influxr_fread_recover}
\title{File fast read and recovery}
\usage{
influxr_fread_recover(
  file,
  sep = "auto",
  nrows = -1L,
  header = "auto",
  skip_lines = 0L,
  select = NULL,
  read_method = "fast",
  text_preprocess_FUN = NULL,
  fill = TRUE,
  strip_extra_whitespace = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{file}{ File name in working directory, path to file (passed through \code{\link[base]{path.expand}} for convenience), or a URL starting http://, file://, etc. Compressed files with extension \file{.gz} and \file{.bz2} are supported if the \code{R.utils} package is installed. }

\item{sep}{ The separator between columns. Defaults to the character in the set \code{[,\\t |;:]} that separates the sample of rows into the most number of lines with the same number of fields. Use \code{NULL} or \code{""} to specify no separator; i.e. each line a single character column like \code{base::readLines} does.}

\item{nrows}{ The maximum number of rows to read. Unlike \code{read.table}, you do not need to set this to an estimate of the number of rows in the file for better speed because that is already automatically determined by \code{fread} almost instantly using the large sample of lines. \code{nrows=0} returns the column names and typed empty columns determined by the large sample; useful for a dry run of a large file or to quickly check format consistency of a set of files before starting to read any of them. }

\item{header}{ Does the first data line contain column names? Defaults according to whether every non-empty field on the first data line is type character. If so, or TRUE is supplied, any empty column names are given a default name. }

\item{skip_lines}{Number of lines to skip.}

\item{select}{ A vector of column names or numbers to keep, drop the rest. \code{select} may specify types too in the same way as \code{colClasses}; i.e., a vector of \code{colname=type} pairs, or a \code{list} of \code{type=col(s)} pairs. In all forms of \code{select}, the order that the columns are specified determines the order of the columns in the result. }

\item{read_method}{Either 'fast', 'R' or a function.
If 'fast' \code{fread} function from package data.table will be
used to read files, it has fast performance but doesn't handle corrupted files well,
the other option 'R' would use internal R \code{readLines} function to read the files
and try to strip irregularities before passing the content again to fread for parsing.
If a function is provided, file content will be read using the provided function.
The function should accept file name as a \code{file} argument. And any additional arguments will be
passed to the function. It should return file content as text which will parsed using fread.
This can be useful when some treatments are necessary for the files before reading e.g. unzip, or format conversion.}

\item{fill}{logical (default is \code{FALSE}). If \code{TRUE} then in case the rows have unequal length, blank fields are implicitly filled.}

\item{verbose}{ Be chatty and report timings? }

\item{...}{Additional arguments passed to \code{fread}}

\item{logical.}{Strip_extra_whitespaces. When TRUE duplicated spaces as
well as trailing and leading white spaces will be removed before processing columns.
Useful when you have the separator as white space to avoid
confusion with the number of columns.}
}
\description{
\code{influxr_fread_recover}.
}
\details{
Fill = FALSE will remove any rows that don't have the same number of columns.
}
