% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-files.R
\name{influxr_list_files}
\alias{influxr_list_files}
\title{List files in a directory}
\usage{
influxr_list_files(
  path = ".",
  pattern = "*",
  pattern_exclude = NULL,
  grep_path = FALSE,
  recursive = FALSE,
  min_size = 0,
  max_size = Inf,
  skip = 0,
  from = NULL,
  to = NULL,
  file_mtime = c("mtime", "filename"),
  mtime_format = NULL,
  nfiles = Inf,
  auto_skip_processed = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{path}{character. Path to the directory.}

\item{pattern}{character. String pattern (regex) used for selecting files for
upload. If \code{pattern} is a character vector of length > 1, its elements
are connected with a logical AND, i.e., all elements must match to select a
file for upload.}

\item{pattern_exclude}{character or \code{NULL}. If character, it is taken as
the filename string pattern used for deselecting files for upload.
If \code{pattern_exclude} has a length > 1, its elements are connected with
a logical OR, i.e., one matching element is sufficient to deselect a file
for upload. If \code{NULL}, no deselection based on filename takes place.}

\item{grep_path}{logical. If \code{TRUE}, the whole file path will be
considered when selecting files based on \code{pattern} and \code{pattern_exclude}.
If \code{FALSE}, only the filename will be considered when selecting files
based on \code{pattern} and \code{pattern_exclude}.}

\item{recursive}{logical. Should the directory be searched recursively?}

\item{min_size}{numeric. Minimum size of files to be selected for upload in bytes.}

\item{max_size}{numeric. Maximum size of files to be selected for upload in bytes.}

\item{skip}{numeric. Number of files to skip. Indexing is done according to
\code{list.files(..., full.names = TRUE)}.}

\item{from}{\code{POSIXt} data time object. If not NULL file list will be
subsetted
based on file modification time (mtime), and files will which have a i
timestamp same as or newer than \code{from} will be kept.}

\item{to}{\code{POSIXt} data time object. If not NULL file list will be subsetted
based on file modification time (mtime), and files will which have a
timestamp older than \code{from} will be kept.}

\item{mtime_format}{Format of modification time.}

\item{nfiles}{integer. Maximum number of files to process.}

\item{auto_skip_processed}{logical. When TRUE \code{influxr_list_files} will
compare the listed files against upload 
status file created by \code{influxr_upload_files} funtion and will skip
files that have been processed and their modification time haven't changed.
To use
this feature \code{store_status_locally} argument has to be enabled when
using \code{influxr_upload_files} function.}

\item{mtime}{logical or character. If \code{TRUE}, data selection is based
on file modification time (mtime).
If character, it is used as the format string for converting the filename
to \code{POSIXt} on which data selection is based.
In both cases, \code{from} and \code{to} specify the data selection time limits.
If \code{FALSE}, data selection is not based on file modification time or
filename timestamp.}

\item{verbose.}{Whether to print output messages.}
}
\description{
\code{influxr_list_files} (recursively) list all files in a directory that
names match criteria.
}
