% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{influxr_select}
\alias{influxr_select}
\title{Selects data points from an influxDB database.}
\usage{
influxr_select(
  client,
  database,
  measurement = "",
  fields = "*",
  from = NULL,
  to = NULL,
  tags = NULL,
  group_by = NULL,
  aggregation = NULL,
  fill = NULL,
  missing = c(-9999, -9999.99, -9999.9999),
  missing_replacement = NA,
  limit = 1e+06,
  desc = FALSE,
  extra_conditions = NULL,
  ...
)
}
\arguments{
\item{client}{Influx_client object created with \code{\link{influxr_connection}}}

\item{database}{character string. Database name.}

\item{measurement}{character string. InfluxDb measurement name.}

\item{fields}{character vector, \code{NULL} or "*". Fields to be 
selected eg. \code{c('T','Humidity', 'w')}. \code{NULL} or '*' will influxr_select all fields.}

\item{from}{POSIX datetime or character in ISO8601 format which will be parsed to date time
eg. '2017-12-31 14:00:00' or simply '2017' for the start of 2017. if \code{NULL} all values will be selected.}

\item{to}{POSIX datetime or character in ISO8601 format which will be parsed to date time
eg. '2017-12-31 14:00:00' or simply '2017' for the start of 2017. if \code{NULL} all values will be selected.
 Note: influxr_select returns data points in the range [from, to[, ie: where time >= from and time < to.}

\item{tags}{named character vector. Only data points tagged with provided
tags will be fetched. eg. \code{c(station = 'Hainich', level = 3, method = 'EC')}.}

\item{group_by}{character string. How data points should be grouped when an aggregation function is specified, 
possible values are \code{'time(1h)'} to aggregate by time, or any other time interval or other field.
 If \code{NULL} no grouping will happen.}

\item{fill}{character string. How should empty datapoints be filled, default to 'null' other 
possible options are \code{"linear", "none", "null" or "previous"}}

\item{limit}{numeric. How many data points should the result be limited to. If \code{NULL} or \code{Inf}
everything will be returned. Default to one million data points.}

\item{desc}{logical, not emplemented yet.}

\item{extra_conditions}{Pass extra coditions to the query.}

\item{...}{passed to \code{influxr_query}. Example: \code {write_to_file = '~/test.csv'}}

\item{aggreagation}{character string. Aggregation function to be used. Currently supported
functions are \code{'MEAN','COUNT','DISTINCT', 'INTEGRAL','MEDIAN','MODE', 'SPREAD','STDDEV','SUM', 'FIRST','LAST', 'MIN', 'MAX'}
if \code{NULL} no aggregation is performed.}
}
\description{
\code{influxr_select} returns the result of a query.
}
\details{
other options that can be passed to the function are \code{ignore_names} and \code{ignore_tags}.
}
\examples{

# Read data from InfluxDB instance and store the result in timed rotating 

\dontrun{

influxr_select(
  influxdb_connection,
  database,
  measurement = 'cpu',
  tags = c(level = 1, station = 'work'),
  verbose = TRUE,
  from = '2017-07-10 12:00:00',
  to = '2017-08-10 12:00:00',
  write_handler = save_timed_rotating_files,
  path = my_path,
  base_file_name = 'cpu_',
  rotation_interval = '1 week',
  trunc_start = TRUE
)
}

}
