% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload-files.R
\name{influxr_upload_files}
\alias{influxr_upload_files}
\title{Upload file(s) to InfluxDB database}
\usage{
influxr_upload_files(
  client,
  files,
  measurement,
  database,
  precision = c("ms", "u", "ns", "s", "m", "h"),
  tags,
  timestamp,
  timestamp_format = "\%Y-\%m-\%d \%H:\%M:\%OS",
  timestamp_offset = 0,
  tz = ifelse(Sys.getenv("TZ") != "", Sys.getenv("TZ"), Sys.timezone()),
  headerline = NULL,
  na.strings = c("", "NA", "N/A", "null", "None"),
  skip_lines = 0L,
  nrows = -1L,
  col.classes = NULL,
  coerce.classes = FALSE,
  select = NULL,
  sep = "auto",
  inclusive = FALSE,
  exclude_timestamp_cols = TRUE,
  recover_files = TRUE,
  read_method = "fast",
  text_preprocess_FUN = NULL,
  upload_preprocess_FUN = NULL,
  stop_on_error = TRUE,
  store_status_locally = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{files}{character vector, containing full file names of files to be uploaded, use 
\code{\link{influxr_list_files}} provided with this package to get a filtered list of files 
according to user specified coditions.}

\item{timestamp}{Where to get time stamp information for the the provided files. Can be one of the following:
\itemize{
\item A numeric vector specifying the column(s) of the files in which the timestamp is contained as a character string.
The string will be converted to \code{POSIXt} according to \code{timestamp_format}, using \code{tz} as the timezone.
If several columns are specified, their contents in each row are pasted together, separated by a single whitespace.
\item A character vector which is converted to \code{POSIXt} format according to \code{format}, using \code{tz} as the timezone.
\item A function to generate timestamp, for instance, when the filename encodes the start
or the end of measurement period. In this case the filename is considered a timestamp 
and converted to \code{POSIXt} format according to \code{timestamp_format}, 
using \code{tz} as the timezone and arguments 
\code{start_of_interval} and \code{resolution} (see there). 
This can be used to process files which do not contain a timestamp variable.
\item A \code{POSIXt} vector.
}}

\item{timestamp_format}{character. Format used for converting the timestamp to \code{POSIXt}}

\item{tz}{character. Timezone of the provided measurements, the timestamp will be converted to the server timestamp before uploading.
The default is the system environment variable "TZ" and if empty the system timezone \code{Sys.timezone()}'.
You can specify other timezones, e.g. for time in Germany, do \code{tz = "Etc/GMT-1"}. 
Please note the sign of the offset, which might be counter intuitive: if you are east of Greenwich you have to give 
a negative offset because you have to decrease your local time to make it become UTC time. 
For timezones west of Greenwich the offset is positive.}

\item{headerline}{numeric or \code{NULL}. If numeric, it is considered the row number of the header line from which column names will be taken. If \code{NULL}, column names will depend on \code{col.names}.}

\item{na.strings}{ A character vector of strings which are to be interpreted as \code{NA} values. By default, \code{",,"} for columns of all types, including type \code{character} is read as \code{NA} for consistency. \code{,"",} is unambiguous and read as an empty string. To read \code{,NA,} as \code{NA}, set \code{na.strings="NA"}. To read \code{,,} as blank string \code{""}, set \code{na.strings=NULL}. When they occur in the file, the strings in \code{na.strings} should not appear quoted since that is how the string literal \code{,"NA",} is distinguished from \code{,NA,}, for example, when \code{na.strings="NA"}. }

\item{skip_lines}{integer. Rows to skip in each file before reading data.}

\item{nrows}{numeric. Number of rows to read. -1 means all lines. See \code{\link{fread}}.}

\item{col.classes}{A character vector of classes (named or unnamed), as \code{\link{read.csv}}. Or a named list of vectors of column names or numbers, see \code{\link{fread}}. If \code{NULL}, column classes are guessed.}

\item{select}{ A vector of column names or numbers to keep, drop the rest. \code{select} may specify types too in the same way as \code{colClasses}; i.e., a vector of \code{colname=type} pairs, or a \code{list} of \code{type=col(s)} pairs. In all forms of \code{select}, the order that the columns are specified determines the order of the columns in the result. }

\item{sep}{character. The separator between columns. See \code{\link{fread}}.}

\item{exclude_timestamp_cols}{logical. Should the column(s) specified in \code{timestamp} be excluded before upload?}

\item{text_preprocess_FUN}{NULL or function. A function to be excuted on the file content before reading as csv}

\item{upload_preprocess_FUN}{NULL or function. A function to be excuted on the content before uploading to the database (eg. calibration)}

\item{stop_on_error}{Whether to stop code excution when a reading error is faced. if True instead of halting
the excution, the file is ignored and a warning is shown at the end.}

\item{verbose}{ Be chatty and report timings? }

\item{...}{Options passed to \code{\link{influxr_fread_recover}}.}

\item{na.string}{character vector. Strings to be replaced with (NA) and hence excluded from uploading to the databases.}

\item{header}{Passed to fread.}

\item{start_of_interval}{logical. Only used if \code{timestamp = "filename"} (see there). 
If \code{TRUE}, the filename is considered to be the first timestamp. 
If \code{FALSE}, the filename is considered the last timestamp.}
}
\description{
\code{influxr_upload_files} uploads a list of files to influxDB.
}
