% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{label_properties}
\alias{label_properties}
\title{Generate label properties from a vector/list}
\usage{
label_properties(values)
}
\arguments{
\item{values}{Either a named vector or named list of property values}
}
\value{
A data structure in the format required by \code{create_label}
}
\description{
Label properties are stored in influx as strings.  Therefore, all elements in the
\code{values} argument should be of length 1 and should be capable of being coerced to a
string if not already a string.
}
\examples{
label_properties(list(p1 = 5, p2 = "ok", p3 = 1))
}
