% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorizations.R
\name{list_authorizations}
\alias{list_authorizations}
\title{List authorizations available on an influxdb server}
\usage{
list_authorizations(con, org = NULL, orgID = NULL, user = NULL, userID = NULL)
}
\arguments{
\item{con}{influxdb connection previously established with \code{connect}}

\item{org}{Only show authorizations that match the organization named
(mutually exclusive with \code{orgID})}

\item{orgID}{Only show authorizations that match the organization id
(mutually exclusive with \code{org})}

\item{user}{Only show authorizations that belong to the user named (mutually
exclusive with \code{userID})}

\item{userID}{Only show authorizations that belong to the userID (mutually
exclusive with \code{user})}
}
\value{
Influx V2 API json response
}
\description{
List authorizations available on an influxdb server
}
\examples{
\dontrun{
con <- connect(url = "https://myserver.mycompany.com:8086",
               org = "my-org",
               token = "my-token")
all_auth <- list_authorizations(con)
org_auth <- list_authorizations(con, org = "my-org")
user_auth <- list_authorizations(con, user = "username")
}
}
