% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buckets.R
\name{list_bucket_owners}
\alias{list_bucket_owners}
\title{List the influx owners associated with the influx bucketID}
\usage{
list_bucket_owners(con, bucketID)
}
\arguments{
\item{con}{influxdb connection previously established with \code{connect}}

\item{bucketID}{id of the bucket whose owners are to be listed}
}
\value{
Influxdb V2 API json response
}
\description{
List the influx owners associated with the influx bucketID
}
