% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buckets.R
\name{list_buckets}
\alias{list_buckets}
\title{Get the list of influx buckets}
\usage{
list_buckets(
  con,
  name = NULL,
  after = NULL,
  id = NULL,
  limit = NULL,
  offset = NULL,
  org = NULL,
  orgID = NULL
)
}
\arguments{
\item{con}{influxdb connection previously established with \code{connect}}

\item{name}{Only return the bucket with the specified name}

\item{after}{The last resource ID from which to seek from.  Use this instead of offset.}

\item{id}{Only return buckets with a specific ID.}

\item{limit}{Integer limit of the number of buckets to return (default 20)}

\item{offset}{Integer >- 0}

\item{org}{The name of the organization}

\item{orgID}{The organization ID}
}
\value{
Influxdb V2 API json response
}
\description{
Get the list of influx buckets
}
