% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{list_labels}
\alias{list_labels}
\title{Get the list of labels available at the influx connection}
\usage{
list_labels(con)
}
\arguments{
\item{con}{influxdb connection previously established with \code{connect}}
}
\value{
influxdb V2 API json response
}
\description{
Get the list of labels available at the influx connection
}
