% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations.R
\name{list_organization_owners}
\alias{list_organization_owners}
\title{List the influx owners associated with the influx orgID}
\usage{
list_organization_owners(con, orgID)
}
\arguments{
\item{con}{influxdb connection previously established with \code{connect()}}

\item{orgID}{id of the organization whose owners are to be listed}
}
\value{
Influxdb V2 API json response
}
\description{
List the influx owners associated with the influx orgID
}
