% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations.R
\name{list_organizations}
\alias{list_organizations}
\title{Get the list of organizations available at the influxdb connection}
\usage{
list_organizations(
  con,
  descending = NULL,
  limit = NULL,
  offset = NULL,
  org = NULL,
  orgID = NULL,
  userID = NULL
)
}
\arguments{
\item{con}{influxdb connection previously established with \code{connect()}}

\item{descending}{boolean}

\item{limit}{integer - default = 20}

\item{offset}{integer >- 0}

\item{org}{Filter organizations specific to the org name}

\item{orgID}{Filter organizations specific to the orgID}

\item{userID}{Fitler organizations to a specific user ID}
}
\value{
Influxdb V2 API json response
}
\description{
Get the list of organizations available at the influxdb connection
}
