% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{list_secrets}
\alias{list_secrets}
\title{List the influxdb secrets associated with an influxdb orgID}
\usage{
list_secrets(con, orgID)
}
\arguments{
\item{con}{influxdb connection previously established with \code{connect}}

\item{orgID}{id of the organization whose secrets are to be listed}
}
\value{
influxdb V2 API json response
}
\description{
List the influxdb secrets associated with an influxdb orgID
}
