% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{list_sources}
\alias{list_sources}
\title{List sources for an influxdb}
\usage{
list_sources(con, org = NULL)
}
\arguments{
\item{con}{influxdb connection previously established with \code{connect}}

\item{org}{string optional organization name}
}
\value{
influxdb V2 API json response
}
\description{
List sources for an influxdb
}
