% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{list_variable_labels}
\alias{list_variable_labels}
\title{List the influx labels associated with the influx variableID}
\usage{
list_variable_labels(con, variableID)
}
\arguments{
\item{con}{the influx connection previously established with \code{connect}}

\item{variableID}{id of the variable whose labels are to be listed}
}
\value{
Influx V2 API json response
}
\description{
List the influx labels associated with the influx variableID
}
