% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{list_variables}
\alias{list_variables}
\title{Get the list of variables available at the influx connection}
\usage{
list_variables(con)
}
\arguments{
\item{con}{the influx connection previously established with \code{connect}}
}
\value{
Influx V2 API json response
}
\description{
Get the list of variables available at the influx connection
}
