% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_tibble.R
\name{map_influx_datatypes}
\alias{map_influx_datatypes}
\title{Convert the datatype annotation of the influx CSV table into
readr::read_csv() datatypes}
\usage{
map_influx_datatypes(header)
}
\arguments{
\item{header}{character. A comma separated list of column data types
as returned by influx's annotated CSV output.}
}
\value{
A column type character string as expected by readr::read_csv()
}
\description{
The first column is always ignored.
}
