% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{MapVariableProperties}
\alias{MapVariableProperties}
\title{Generate a MapVariableProperties variable type from a named vector or named list}
\usage{
MapVariableProperties(values)
}
\arguments{
\item{values}{Either a named vector or named list of property values}
}
\value{
A data structure in the format required by \code{create_variable} to create
a MapVariableProperties type variable
}
\description{
Map variable properties are stored in influx as strings.  Therefore, all elements in the
\code{values} argument should be of length 1 and should be capable of being coerced to a
string if not already a string.
}
\examples{
MapVariableProperties(c(p1 = "v1", p2 = "v2", p3 = "v3"))
MapVariableProperties(list(p1 = "v1", p2 = 1.0, p3 = 5))
}
