% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorizations.R
\name{permissions}
\alias{permissions}
\title{Create a set of read and write permissions using the list of resources provided}
\usage{
permissions(read = NULL, write = NULL, read_write = NULL)
}
\arguments{
\item{read}{list of resources to create read permissions for}

\item{write}{list of resources to create write permissions for}

\item{read_write}{list of resources to create both read and write permissions for}
}
\value{
permission data structure used when creating an authorization
with \code{create_authorization}
}
\description{
Create a set of read and write permissions using the list of resources provided
}
\examples{
permissions(read_write = resource(type = "buckets"))
permissions(read = list(resource(type = "buckets"),
                        resource(type = "authorizations", org = "myorg")))
}
