% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{post_line_protocol}
\alias{post_line_protocol}
\title{Post the line protocol formatted data to the connection}
\usage{
post_line_protocol(con, bucket, data, precision)
}
\arguments{
\item{con}{the influx database endpoint created by \code{connect}}

\item{bucket}{the name of the influx bucket to write the data to}

\item{data}{The data (string or character vector) to be posted as "text/plain; charset=utf-8"}

\item{precision}{Precision of the time point being written.  This value is any valid
influx precision value.}
}
\value{
Influx V2 API \code{httr::POST} response
}
\description{
Post the line protocol formatted data to the connection
}
