% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{print.influxdb}
\alias{print.influxdb}
\title{Formatter for printing the influxdb class}
\usage{
\method{print}{influxdb}(x, ...)
}
\arguments{
\item{x}{The influxdb object passed to print}

\item{...}{ignored}
}
\value{
the object as invisible
}
\description{
Formatter for printing the influxdb class
}
\examples{
\dontrun{
con <- connect(url = "https://myserver.mycompany.com:8086",
               org = "my-org",
               token = "my-token")
print(con)
}
}
