% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_tibble.R
\name{query_tibble}
\alias{query_tibble}
\title{Send a query command to an influx server and parse the results as a list of
tibbles}
\usage{
query_tibble(con, code, flatten = TRUE, max_tries = 1, debug = FALSE)
}
\arguments{
\item{con}{influx database connection returned by \code{connect}}

\item{code}{a string containing the flux code to run as the query}

\item{flatten}{a boolean indicating that a single tibble should be returned instead
of a list when only a single table is returned.}

\item{max_tries}{numeric indicating how many times the request should be attempted}

\item{debug}{print verbose messages to standard output}
}
\value{
A tibble for each CSV table returned.
}
\description{
If the http command returns an error, an error is thrown.
}
\examples{
\dontrun{
con <- connect(url = "http://myserver.mycompany.com",
               org = "my-org",
               token = "my-token")
code <- '
import "influxdata/influxdb/schema"
schema.tagKeys(
  bucket: "my-bucket",
  predicate: (r) => true,
  start: 2021-03-01-T00:00:00Z)
'
df <- query_tibble(con, code)
}
}
