% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{QueryVariableProperties}
\alias{QueryVariableProperties}
\title{Generate a QueryVariableProperties variable from a query string}
\usage{
QueryVariableProperties(query, language = "flux")
}
\arguments{
\item{query}{string query}

\item{language}{defaults to "flux"}
}
\value{
A data structure in the format required by \code{create_variable} to create
a QueryVariableProperties variable
}
\description{
Generate a QueryVariableProperties variable from a query string
}
\examples{

QueryVariableProperties('schema.TagKeys("bucket-name")')
}
