% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{quote_specials}
\alias{quote_specials}
\title{Quote characters following influx quoting rules for line protocol items}
\usage{
quote_specials(
  vec,
  purpose = c("measurement", "tag_key", "tag_value", "field_key", "field_value")
)
}
\arguments{
\item{vec}{vector of data to quote each element of}

\item{purpose}{the purpose of the vector of data in the line protocol output.  This purpose
defines the influx quoting rules used.}
}
\value{
vector data converted to character strings and quoted according to its purpose
}
\description{
Quote characters following influx quoting rules for line protocol items
}
\examples{
quote_specials("example 1", purpose = "measurement")
quote_specials(c("name=1", "name2", "name,3"), purpose = "tag_key")
}
