% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health.R
\name{ready}
\alias{ready}
\title{Get the readiness of an instance at startup}
\usage{
ready(con)
}
\arguments{
\item{con}{influxdb connection previously established with \code{connect}}
}
\value{
Influxdb API V2.0 json response
}
\description{
Get the readiness of an instance at startup
}
