% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buckets.R
\name{remove_member_from_bucket}
\alias{remove_member_from_bucket}
\title{Delete the association of an influx memberID with an influx bucketID}
\usage{
remove_member_from_bucket(con, userID, bucketID)
}
\arguments{
\item{con}{influxdb connection previously established with \code{connect}}

\item{userID}{id of the member to disassociate}

\item{bucketID}{id of the bucket to disassociate}
}
\value{
Influxdb V2 API http response
}
\description{
Delete the association of an influx memberID with an influx bucketID
}
