% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations.R
\name{remove_owner_from_organization}
\alias{remove_owner_from_organization}
\title{Delete the association of an influx ownerID with an influx orgID}
\usage{
remove_owner_from_organization(con, userID, orgID)
}
\arguments{
\item{con}{influxdb connection previously established with \code{connect()}}

\item{userID}{id of the owner to disassociate}

\item{orgID}{id of the organization to disassociate}
}
\value{
Influxdb V2 API response
}
\description{
Delete the association of an influx ownerID with an influx orgID
}
