% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{request}
\alias{request}
\title{Build a basic http request to an influxdb server}
\usage{
request(con, path = NULL, query = NULL, method = NULL)
}
\arguments{
\item{con}{An influxdb connection previous established using connect()}

\item{path}{Optional string path to append to the base url}

\item{query}{Optional list of query name/value pairs to append to the url}

\item{method}{Set a custom HTTP method like HEAD, DELETE, PATCH, UPDATE, or
OPTIONS. The default method is GET for requests without a body, and POST
for requests with a body.}
}
\value{
An httr2 request object
}
\description{
Additional modifications can be made to the returned object before executing
the query with httr2::req_perform().
}
