% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorizations.R
\name{resource}
\alias{resource}
\title{Return a resource data structure in required when creating permissions}
\usage{
resource(
  type = influxdb_resource_types(),
  id = NULL,
  name = NULL,
  org = NULL,
  orgID = NULL
)
}
\arguments{
\item{type}{the type of influx resource the permission will be applied to.
\code{influx_resource_types()} returns a complete list of available resource
types}

\item{id}{If id set, the permission applies to a specific resource.  Otherwise it applies
to all resources of that type}

\item{name}{Optional name of the resource if it has a name}

\item{org}{Optional name of the organization with the orgID}

\item{orgID}{If orgID is set, the permission is for all resources owned by that org.
If not, it is a permission for all resources of that resource type}
}
\value{
data structure representing a single resource.  The resource must still be combined
with an action ("read" or "write") to create a permission.
}
\description{
Return a resource data structure in required when creating permissions
}
\examples{
resource(type = "buckets")
resource(type = "authorizations", org = "myorg")
}
