% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{respond}
\alias{respond}
\title{Retrieve an influxdb server http response to a basic query}
\usage{
respond(con, path = NULL, query = NULL, method = NULL)
}
\arguments{
\item{con}{An influxdb connection previous established using connect()}

\item{path}{Optional string path to append to the base url}

\item{query}{Optional list of query name/value pairs to append to the url}

\item{method}{Set a custom HTTP method like HEAD, DELETE, PATCH, UPDATE, or
OPTIONS. The default method is GET for requests without a body, and POST
for requests with a body.}
}
\value{
The httr2 formatted response from the server
}
\description{
Retrieve an influxdb server http response to a basic query
}
