% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buckets.R
\name{retentionRules}
\alias{retentionRules}
\title{Create a retention rule for a bucket}
\usage{
retentionRules(
  type = "expire",
  everySeconds = 0L,
  shardGroupDurationSeconds = NULL
)
}
\arguments{
\item{type}{default "expire"}

\item{everySeconds}{>= 0; 0 means infinite}

\item{shardGroupDurationSeconds}{Shard duration measured in seconds. Optional.}
}
\value{
a retention rule data structure
}
\description{
Create a retention rule for a bucket
}
\examples{
retentionRules()
retentionRules(everySeconds = 60L * 60L * 24L * 7L) # weekly
}
