% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buckets.R
\name{retrieve_bucket}
\alias{retrieve_bucket}
\title{Retrieve a bucket by bucketID from the influx database}
\usage{
retrieve_bucket(con, bucketID)
}
\arguments{
\item{con}{influxdb connection previously established with \code{connect}}

\item{bucketID}{the ID of the bucket to retrieve}
}
\value{
Influxdb V2 API json response
}
\description{
Retrieve a bucket by bucketID from the influx database
}
