% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{retrieve_current_user}
\alias{retrieve_current_user}
\title{Retrieve the currently authenticated influxdb user}
\usage{
retrieve_current_user(con)
}
\arguments{
\item{con}{the influx connection previously established with \code{connect}}
}
\value{
Influx V2 API \code{httr::GET} response
}
\description{
Retrieve the currently authenticated influxdb user
}
