% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{retrieve_feature_flags}
\alias{retrieve_feature_flags}
\title{Retrieve the feature flags for the currently authenticated influxdb user}
\usage{
retrieve_feature_flags(con)
}
\arguments{
\item{con}{the influx connection previously established with \code{connect}}
}
\value{
Influx V2 API \code{httr::GET} response
}
\description{
Retrieve the feature flags for the currently authenticated influxdb user
}
