% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{retrieve_label}
\alias{retrieve_label}
\title{Retrieve a label by labelID from the influx database}
\usage{
retrieve_label(con, labelID)
}
\arguments{
\item{con}{influxdb connection previously established with \code{connect}}

\item{labelID}{the ID of the label to retrieve}
}
\value{
influxdb V2 API json response
}
\description{
Retrieve a label by labelID from the influx database
}
