% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations.R
\name{retrieve_organization}
\alias{retrieve_organization}
\title{Retrieve a organization by orgID from the influx database}
\usage{
retrieve_organization(con, orgID)
}
\arguments{
\item{con}{influxdb connection previously established with \code{connect()}}

\item{orgID}{the ID of the organization to retrieve}
}
\value{
Influxdb V2 API json response
}
\description{
Retrieve a organization by orgID from the influx database
}
