% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{retrieve_source}
\alias{retrieve_source}
\title{Retrieve an influxdb source}
\usage{
retrieve_source(con, sourceID)
}
\arguments{
\item{con}{influxdb connection previously established with \code{connect}}

\item{sourceID}{string source identifier}
}
\value{
influxdb V2 API http response
}
\description{
Retrieve an influxdb source
}
