% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{retrieve_variable}
\alias{retrieve_variable}
\title{Retrieve a variable by variableID from the influx database}
\usage{
retrieve_variable(con, variableID)
}
\arguments{
\item{con}{the influx connection previously established with \code{connect}}

\item{variableID}{the ID of the variable to retrieve}
}
\value{
Influx V2 API json response
}
\description{
Retrieve a variable by variableID from the influx database
}
