% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timed-rotating-save.R
\name{save_timed_rotating_files}
\alias{save_timed_rotating_files}
\title{Save into timed rotating files.}
\usage{
save_timed_rotating_files(
  x,
  path,
  base_file_name,
  rotation_interval = "1 day",
  time_suffix_format = "auto",
  trunc_start = TRUE,
  chunk_number = NULL,
  drop_time = FALSE,
  append = TRUE,
  filter = NULL,
  ...
)
}
\arguments{
\item{x}{data.table. Data to be written, first variable should be named time and of type POSIXct}

\item{path}{string. Path where files should be saved.}

\item{base_file_name}{string. File name to be used as a base.}

\item{rotation_interval}{Duration of data to be written on each file eg. '1 day' or '1 hour 30 minutes'}

\item{time_suffix_format}{Time format for file name suffix eg. '%Y%m%d%H%M%S' if left to auto, it will automatically choose formatting.}

\item{trunc_start}{Whether to truncate the start time to the nearest rotation_interval.}

\item{chunk_number}{numeric. Number of the chunk.}

\item{append}{logical. Whether to append if same file exists on disk}

\item{...}{passed to fwrite}
}
\description{
Save into timed rotating files.
}
\details{
Important: date time columns are converted to UTC upon saving.
}
