% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorizations.R
\name{update_authorization}
\alias{update_authorization}
\title{Update an influxdb authorization}
\usage{
update_authorization(con, authID, description = NULL, status = NULL)
}
\arguments{
\item{con}{influxdb connection previously established with \code{connect}}

\item{authID}{the ID of the authorization to retrieve}

\item{description}{String description of the authorization}

\item{status}{String "active" or "inactive"}
}
\value{
Influx V2 API json response
}
\description{
Update an influxdb authorization
}
