% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations.R
\name{update_organization}
\alias{update_organization}
\title{Update a organization on the influx server by orgID}
\usage{
update_organization(con, orgID, name = NULL, description = NULL)
}
\arguments{
\item{con}{influxdb connection previously established with \code{connect()}}

\item{orgID}{the ID of the organization to update}

\item{name}{new name of the organization if set}

\item{description}{new description of the organization if set}
}
\value{
Influxdb V2 API json response
}
\description{
Update a organization on the influx server by orgID
}
