% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{update_password}
\alias{update_password}
\title{Update the password for the current influxdb user}
\usage{
update_password(con, password)
}
\arguments{
\item{con}{the influx connection previously established with \code{connect}}

\item{password}{new password string}
}
\value{
Influx V2 API \code{httr::PUT} response
}
\description{
Update the password for the current influxdb user
}
