% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{update_secrets}
\alias{update_secrets}
\title{Associate an influx user with an influx orgID}
\usage{
update_secrets(con, orgID, secrets)
}
\arguments{
\item{con}{influxdb connection previously established with \code{connect}}

\item{orgID}{id of the organization with which to associate the secret}

\item{secrets}{list of key/value pairs}
}
\value{
influxdb V2 API json response
}
\description{
Associate an influx user with an influx orgID
}
