% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{update_source}
\alias{update_source}
\title{Update an influxdb source}
\usage{
update_source(
  con,
  sourceID,
  default = NULL,
  defaultRP = NULL,
  id = NULL,
  insecureSkipVerify = NULL,
  links = NULL,
  metaUrl = NULL,
  name = NULL,
  orgID = NULL,
  password = NULL,
  sharedSecret = NULL,
  telegraf = NULL,
  token = NULL,
  type = NULL,
  url = NULL,
  username = NULL
)
}
\arguments{
\item{con}{influxdb connection previously established with \code{connect}}

\item{sourceID}{string source identifier}

\item{default}{boolean}

\item{defaultRP}{string}

\item{id}{string}

\item{insecureSkipVerify}{boolean}

\item{links}{list with named strings "buckets", "health", "query", "self"}

\item{metaUrl}{string <uri>}

\item{name}{string}

\item{orgID}{string}

\item{password}{string}

\item{sharedSecret}{string}

\item{telegraf}{string}

\item{token}{string}

\item{type}{string}

\item{url}{string <uri>}

\item{username}{string}
}
\value{
influxdb V2 API json response
}
\description{
Update an influxdb source
}
