% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{update_user}
\alias{update_user}
\title{Update an influxdb user}
\usage{
update_user(
  con,
  userID,
  name,
  oauthID = NULL,
  status = c("active", "inactive")
)
}
\arguments{
\item{con}{the influx connection previously established with \code{connect}}

\item{userID}{string ID of the user to update}

\item{name}{string name of the user}

\item{oauthID}{string}

\item{status}{string either "active" or "inactive"}
}
\value{
Influx V2 API \code{httr::POST} response
}
\description{
Update an influxdb user
}
