% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{update_user_password}
\alias{update_user_password}
\title{Update an influxdb user password}
\usage{
update_user_password(con, userID, password)
}
\arguments{
\item{con}{the influx connection previously established with \code{connect}}

\item{userID}{id of the user}

\item{password}{new password for the user}
}
\value{
Influx V2 API \code{httr::POST} response
}
\description{
Update an influxdb user password
}
