% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{variable_labels}
\alias{variable_labels}
\title{Generate a list of labels to be associated with a variable}
\usage{
variable_labels(values)
}
\arguments{
\item{values}{a named list of \code{label_properties}}
}
\value{
a data structure suitable as input to the \code{labels} argument of
\code{create_variable}
}
\description{
Generate a list of labels to be associated with a variable
}
\examples{
labels <- list(l1 = label_properties(list(p1 = 1, p2 = 2, p3 = "3")),
               l2 = label_properties(list(p4 = "ok", p5 = "10")),
               l3 = label_properties(list(p6 = "bucket-name")))
variable_labels(labels)
}
