% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write}
\alias{write}
\title{Write a dataframe to a influx connection}
\usage{
write(
  con,
  data,
  bucket,
  measurement = NULL,
  measurement_col = NULL,
  time_col = "time",
  tag_cols = NULL,
  field_cols = NULL,
  batch_size = 5000,
  precision = "s"
)
}
\arguments{
\item{con}{the influx database endpoint created by \code{connect}}

\item{data}{a dataframe containing the data to write}

\item{bucket}{The name of the bucket to write the data to.}

\item{measurement}{the name of the influx measurement to use.  Cannot be specified if
\code{measurement_col} is also specified.}

\item{measurement_col}{the name of the column in the data frame that contains the measurement
name to use for each row of data.  Cannot be specified if \code{measurement} is also specified.}

\item{time_col}{The name of the column containing the timepoint for each row.}

\item{tag_cols}{The names of the columns that are to be treated as tags by influx}

\item{field_cols}{The names of the columns that are to be treated as fields by influx}

\item{batch_size}{The number of lines to write per HTTP POST.  5000 is the optimal number for
influx line protocol writes.}

\item{precision}{The precision of the data to write.}
}
\value{
Influx V2 API \code{httr::POST} response of the last batch of data that was written
}
\description{
Write a dataframe to a influx connection
}
