% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{write_csv_handler}
\alias{write_csv_handler}
\title{write datatable chunk to file}
\usage{
write_csv_handler(x, file_name, chunk_number, append = TRUE)
}
\arguments{
\item{x}{datatable chunk.}

\item{file_name}{full file name.}

\item{chunk_number}{chunk number in order to manage header.}

\item{append}{logical. Whether to append to existing file.}
}
\description{
write datatable chunk to file
}
\examples{
# Write `influxr_select` output to a csv file

\dontrun{
output_file <- tempfile()
influxr_select(tss_client, database = 'metrics', measurement = 'cpu',
               tags = c(level = 1, station = 'home'),
               from = '2017-07-10 12:00:00', to = '2017-07-20 12:00:00',
               write_handler = write_csv_handler,
               file_name = output_file, append = TRUE)
}
}
